refactor(webui): 重构视频脚本生成-目录结果
- 将视频脚本生成相关代码从 script_settings.py 移动到新的 generate_script_docu.py 文件
- 新增 base.py 文件，提取公共工具函数
- 优化代码结构，提高可维护性和可读性- 重构函数名称，更清晰地反映功能
